+++ includes/getsettings.php
@@ -15,6 +15,7 @@ if ($settings) {
     $settings['convertCurrency'] = $settings['convert_currency'] ? 'true': 'false';
     $settings['removeBackground'] = $settings['remove_background'] ? 'true': 'false';
     $settings['hideDisabledSubscriptions'] = $settings['hide_disabled'] ? 'true': 'false';
+    $settings['archiveAfterdays'] = $settings['archive_afterdays'] ? $settings['archive_afterdays'] : 0;
 }
 
 $query = "SELECT * FROM custom_colors WHERE user_id = :userId";